+++ io_annocfg/anno_objects.py
@@ -14,1459 +14,16 @@ from bpy.props import EnumProperty, BoolProperty, PointerProperty, IntProperty, 
 from bpy.types import PropertyGroup, Panel, Operator, UIList
 import bmesh
 import sys
-def str_to_class(classname):
-    return getattr(sys.modules[__name__], classname)
+
 from collections import defaultdict
 from math import radians
 from .prefs import IO_AnnocfgPreferences
-from .utils import data_path_to_absolute_path, to_data_path
+from .utils import *
+from .transform import Transform
+from .material import Material, ClothMaterial
 from .feedback_ui import FeedbackConfigItem, GUIDVariationListItem, FeedbackSequenceListItem
 from . import feedback_enums
 
-def parse_float_node(node, query, default_value = 0.0):
-    value = default_value
-    if node.find(query) is not None:
-        value = float(node.find(query).text)
-    return value
-
-def get_float(node, query, default_value = 0.0):
-    value = default_value
-    if node.find(query) is not None:
-        value = float(node.find(query).text)
-    return value
-
-
-
-class Transform:
-    """
-    Parses an xml tree node for transform operations, stores them and can apply them to a blender object.
-    """
-    def __init__(self, loc = [0,0,0], rot = [1,0,0,0], sca = [1,1,1], anno_coords = True):
-        self.location = loc
-        self.rotation = rot
-        self.rotation_euler = [0,0,0]
-        self.euler_rotation = False
-        self.scale = sca
-        self.anno_coords = anno_coords
-        
-    def get_component_value(self, component_name: str) -> float:
-        component_by_name = {
-            "location.x": self.location[0],
-            "location.y": self.location[1],
-            "location.z": self.location[2],
-            
-            "rotation.w": self.rotation[0],
-            "rotation.x": self.rotation[1],
-            "rotation.y": self.rotation[2],
-            "rotation.z": self.rotation[3],
-            
-            "rotation_euler.x": self.rotation_euler[0],
-            "rotation_euler.y": self.rotation_euler[1],
-            "rotation_euler.z": self.rotation_euler[2],
-            
-            "scale.x": self.scale[0],
-            "scale.y": self.scale[1],
-            "scale.z": self.scale[2],
-        }
-        return component_by_name[component_name]
-        
-    def get_component_from_node(self, node: ET.Element, transform_paths: Dict[str, str], component: str, default = 0.0) -> float:
-        query = transform_paths.get(component, None)
-        if not query:
-            return default
-        value = float(get_text_and_delete(node, query, str(default)))
-        return value
-        
-    @classmethod
-    def from_node(cls, node: ET.Element, transform_paths, enforce_equal_scale: bool, euler_rotation: bool = False) -> Transform:
-        instance = cls()
-        instance.location[0] = instance.get_component_from_node(node, transform_paths, "location.x")
-        instance.location[1] = instance.get_component_from_node(node, transform_paths, "location.y")
-        instance.location[2] = instance.get_component_from_node(node, transform_paths, "location.z")
-        
-        instance.scale[0] = instance.get_component_from_node(node, transform_paths, "scale.x", 1.0)
-        instance.scale[1] = instance.get_component_from_node(node, transform_paths, "scale.y", 1.0)
-        instance.scale[2] = instance.get_component_from_node(node, transform_paths, "scale.z", 1.0)
-        if enforce_equal_scale:
-            instance.scale[1] = instance.scale[0]
-            instance.scale[2] = instance.scale[1]
-        
-        if not euler_rotation:        
-            instance.rotation[0] = instance.get_component_from_node(node, transform_paths, "rotation.w", 1.0)
-            instance.rotation[1] = instance.get_component_from_node(node, transform_paths, "rotation.x")
-            instance.rotation[2] = instance.get_component_from_node(node, transform_paths, "rotation.y")
-            instance.rotation[3] = instance.get_component_from_node(node, transform_paths, "rotation.z")
-        else:
-            instance.rotation_euler[0] = instance.get_component_from_node(node, transform_paths, "rotation_euler.x")
-            instance.rotation_euler[1] = instance.get_component_from_node(node, transform_paths, "rotation_euler.y")
-            instance.rotation_euler[2] = instance.get_component_from_node(node, transform_paths, "rotation_euler.z")
-            instance.euler_rotation = True
-            
-        instance.anno_coords = True
-        return instance
-        
-    @classmethod
-    def from_blender_object(cls, obj, enforce_equal_scale: bool, euler_rotation: bool = False) -> Transform:
-        if enforce_equal_scale:
-            if len(set([obj.scale[0], obj.scale[1], obj.scale[2]])) > 1:
-                print(obj.name, "Cannot have different scale values on xyz")
-        instance = cls(obj.location, [1,0,0,0], obj.scale, False)
-        if not euler_rotation:
-            obj.rotation_mode = "QUATERNION"
-            instance.rotation = list(obj.rotation_quaternion)
-            return instance
-        obj.rotation_mode = "XYZ"
-        instance.rotation_euler = [obj.rotation_euler.x, obj.rotation_euler.y, obj.rotation_euler.z]
-        return instance
-    
-    def convert_to_blender_coords(self):
-        if not self.anno_coords:
-            return
-        if IO_AnnocfgPreferences.mirror_models():
-            self.location = (-self.location[0], -self.location[2], self.location[1])
-            self.rotation = (self.rotation[0], self.rotation[1], self.rotation[3], -self.rotation[2])
-        else:     
-            self.location = (self.location[0], -self.location[2], self.location[1])
-            self.rotation = (self.rotation[0], self.rotation[1], self.rotation[3], self.rotation[2])
-        self.rotation_euler = (self.rotation_euler[0], self.rotation_euler[2], self.rotation_euler[1])
-        self.scale = (self.scale[0], self.scale[2], self.scale[1])
-        
-        self.anno_coords = False
-
-    def convert_to_anno_coords(self):
-        if self.anno_coords:
-            return
-        if IO_AnnocfgPreferences.mirror_models():
-            self.location = (-self.location[0], self.location[2], -self.location[1])
-            self.rotation = (self.rotation[0], self.rotation[1], -self.rotation[3], self.rotation[2])
-        else:     
-            self.location = (self.location[0], self.location[2], -self.location[1])
-            self.rotation = (self.rotation[0], self.rotation[1], self.rotation[3], self.rotation[2])
-        self.rotation_euler = (self.rotation_euler[0], self.rotation_euler[2], self.rotation_euler[1])
-        self.scale = (self.scale[0], self.scale[2], self.scale[1])
-        
-        self.anno_coords
-    
-    def mirror_mesh(self, object):
-        if not IO_AnnocfgPreferences.mirror_models():
-            return
-        if not object.data or not hasattr(object.data, "vertices"):
-            return
-        for v in object.data.vertices:
-            v.co.x *= -1.0
-            
-        #Inverting normals for import AND Export they are wrong because of scaling on the x axis.
-        #Warn people that this will break exports from .blend files made with an earlier version!!!
-        mesh = object.data
-        bm = bmesh.new()
-        bm.from_mesh(mesh) # load bmesh
-        for f in bm.faces:
-             f.normal_flip()
-        bm.normal_update() # not sure if req'd
-        bm.to_mesh(mesh)
-        mesh.update()
-        bm.clear() #.. clear before load next
-    
-    def apply_to(self, object):
-        if self.anno_coords:
-            self.convert_to_blender_coords()
-        self.mirror_mesh(object)
-        object.location = self.location
-        if not self.euler_rotation:
-            object.rotation_mode = "QUATERNION"
-            object.rotation_quaternion = self.rotation
-        else:
-            object.rotation_mode = "XYZ"
-            object.rotation_euler = self.rotation_euler
-        object.scale = self.scale
-
-class BoolPropertyGroup(PropertyGroup):
-    tag : StringProperty(name = "", default = "SomeBool")
-    value : BoolProperty(name = "", default = False)
-
-class FeedbackSequencePropertyGroup(PropertyGroup):
-    tag : StringProperty(name = "", default = "SomeSequence")
-    value : EnumProperty(
-        name='',
-        description='Animation Sequence',
-        items= feedback_enums.animation_sequences,
-        default='idle01'
-    )
-
-class IntPropertyGroup(PropertyGroup):
-    tag : StringProperty(name = "", default = "SomeInt")
-    value : IntProperty(name = "", default = 0)
-class StringPropertyGroup(PropertyGroup):
-    tag : StringProperty(name = "", default = "SomeString")
-    value : StringProperty(name = "", default = "")
-
-class FilenamePropertyGroup(PropertyGroup):
-    tag : StringProperty(name = "", default = "SomeString")
-    value : StringProperty(name = "", default = "", subtype = "FILE_PATH")
-class FloatPropertyGroup(PropertyGroup):
-    tag : StringProperty(name = "", default = "SomeFloat")
-    value : FloatProperty(name = "", default = 0.0)
-    
-class ColorPropertyGroup(PropertyGroup):
-    tag : StringProperty(name = "", default = "SomeColor")
-    value : FloatVectorProperty(name = "", default = [0.0, 0.0, 0.0], subtype = "COLOR", min= 0.0, max = 1.0)
-
-class ObjectPointerPropertyGroup(PropertyGroup):
-    tag : StringProperty(name = "", default = "SomeObject")
-    value : PointerProperty(name = "", type= bpy.types.Object)
-
-
-class Converter(ABC):
-    @classmethod
-    @abstractmethod
-    def data_type(cls):
-        pass
-    @classmethod
-    def from_string(cls, s):
-        """Convert the string s from the input xml node into a blender representation of type data_type()
-
-        Args:
-            s (str): xml_node.text
-
-        Returns:
-            data_type(): Blender representation
-        """
-        value = cls.data_type()
-        try:
-            value = cls.data_type()(s)
-        except:
-            print(f"Error: failed to convert {s} to {cls.data_type()}")
-        return cls.data_type()(s)
-    @classmethod
-    def to_string(cls, value):
-        """Convert the blender representation value into a string for the xml node.
-        Args:
-            value (daty_type()): Blender representation
-
-        Returns:
-            str: XML text string
-        """  
-        return str(value)
-
-class StringConverter(Converter):
-    @classmethod
-    def data_type(cls):
-        return str
-
-class BoolConverter(Converter):
-    @classmethod
-    def data_type(cls):
-        return bool
-    @classmethod
-    def from_string(cls, s):
-        return bool(int(s))
-    @classmethod
-    def to_string(cls, value):
-        return str(int(value))
-        
-class IntConverter(Converter):
-    @classmethod
-    def data_type(cls):
-        return int
-
-class FloatConverter(Converter):
-    @classmethod
-    def data_type(cls):
-        return float
-    @classmethod
-    def to_string(cls, value):
-        return format_float(value)
-class FeedbackSequenceConverter(Converter):
-    @classmethod
-    def data_type(cls):
-        return string
-    @classmethod
-    def from_string(cls, s): 
-        seq_id = int(s)
-        return feedback_enums.NAME_BY_SEQUENCE_ID.get(seq_id, "none")
-    @classmethod
-    def to_string(cls, value): 
-        seq_id = feedback_enums.SEQUENCE_ID_BY_NAME.get(value, -1)
-        return str(seq_id)
-    
-class ObjectPointerConverter(Converter):
-    @classmethod
-    def data_type(cls):
-        return string
-    @classmethod
-    def from_string(cls, s): 
-        return bpy.data.objects[s]
-    @classmethod
-    def to_string(cls, value): 
-        if value is None:
-            return ""
-        return value.name
-
-
-class ColorConverter(Converter):
-    @classmethod
-    def data_type(cls):
-        return str
-    @classmethod
-    def from_string(cls, s): #f.e. COLOR[1.0, 0.5, 0.3]
-        values = s.replace(" ", "").replace("_COLOR[", "").replace("]", "").split(",")
-        assert len(values) == 3
-        return [format_float(value) for value in values]
-    @classmethod
-    def to_string(cls, value): #f.e. [1.0, 0.5, 0.3]
-        assert len(value) == 3
-        return f"_COLOR[{', '.join([str(val) for val in value])}]"
-
-converter_by_tag = {
-    "ConfigType": StringConverter,
-    "FileName" : StringConverter,
-    "Name" : StringConverter,
-    "AdaptTerrainHeight" : BoolConverter,
-    "HeightAdaptationMode" : BoolConverter,
-    "DIFFUSE_ENABLED" : BoolConverter,
-    "NORMAL_ENABLED" : BoolConverter,
-    "METALLIC_TEX_ENABLED" : BoolConverter,
-    "SEPARATE_AO_TEXTURE" : BoolConverter, 
-    "HEIGHT_MAP_ENABLED" : BoolConverter,
-    "NIGHT_GLOW_ENABLED" : BoolConverter,
-    "DYE_MASK_ENABLED" : BoolConverter,
-    "cUseTerrainTinting" : BoolConverter,
-    "SELF_SHADOWING_ENABLED" : BoolConverter, 
-    "WATER_CUTOUT_ENABLED" : BoolConverter,
-    "ADJUST_TO_TERRAIN_HEIGHT" : BoolConverter, 
-    "GLOW_ENABLED": BoolConverter,
-    "SequenceID": FeedbackSequenceConverter,
-    "m_IdleSequenceID": FeedbackSequenceConverter,
-    "BlenderModelID": ObjectPointerConverter,
-    "BlenderParticleID": ObjectPointerConverter,
-}
-
-def get_converter_for(tag, value_string):
-    if tag in converter_by_tag:
-        return converter_by_tag[tag]
-    if value_string.startswith("_COLOR["):
-        return ColorConverter
-    if value_string.isnumeric() or value_string.lstrip("-").isnumeric():
-        return IntConverter
-    if is_type(float, value_string):
-        return FloatConverter
-
-    #TODO: CDATA Converter, mIdleSequenceConverter, etc
-    return StringConverter
-
-class XMLPropertyGroup(PropertyGroup):
-    tag : StringProperty(name = "", default = "")
-    
-    config_type : StringProperty(name = "", default = "")
-    
-    feedback_sequence_properties : CollectionProperty(name = "FeedbackSequences", type = FeedbackSequencePropertyGroup)
-    boolean_properties : CollectionProperty(name = "Bools", type = BoolPropertyGroup)
-    filename_properties : CollectionProperty(name = "Filenames", type = FilenamePropertyGroup)
-    string_properties : CollectionProperty(name = "Strings", type = StringPropertyGroup)
-    int_properties : CollectionProperty(name = "Ints", type = IntPropertyGroup)
-    float_properties : CollectionProperty(name = "Floats", type = FloatPropertyGroup)
-    color_properties : CollectionProperty(name = "Colors", type = ColorPropertyGroup)
-    object_pointer_properties : CollectionProperty(name = "Objects", type = ObjectPointerPropertyGroup)
-    dynamic_properties : CollectionProperty(name = "DynamicProperties", type = XMLPropertyGroup)
-    
-    
-    hidden : BoolProperty(name = "Hide", default = False)
-    deleted : BoolProperty(name = "Delete", default = False)
-    
-    def reset(self):
-        self.feedback_sequence_properties.clear()
-        self.boolean_properties.clear()
-        self.filename_properties.clear()
-        self.string_properties.clear()
-        self.int_properties.clear()
-        self.float_properties.clear()
-        self.color_properties.clear()
-        self.object_pointer_properties.clear()
-        self.dynamic_properties.clear()
-        self.deleted = False
-        self.hidden = False
-    
-    def remove(self, tag):
-        for container in [self.feedback_sequence_properties, 
-                          self.boolean_properties,
-                          self.filename_properties,
-                          self.string_properties,
-                          self.int_properties,
-                          self.float_properties,
-                          self.color_properties,
-                          self.dynamic_properties]:
-            for i,prop in enumerate(container):
-                if prop.tag == tag:
-                    container.remove(i)
-                    return True
-        return False
-    def get_string(self, tag, default = None):
-        for item in self.string_properties:
-            if item.tag == tag:
-                return item.value
-        for item in self.filename_properties:
-            if item.tag == tag:
-                return item.value
-        return default
-    def set(self, tag, value_string, replace = False):
-        converter = get_converter_for(tag, value_string)
-        value = converter.from_string(value_string)
-        
-        # Special fields
-        if tag == "ConfigType":
-            self.config_type = value
-            return
-            
-        properties_by_converter = {
-            BoolConverter: self.boolean_properties,
-            StringConverter: self.string_properties,
-            IntConverter: self.int_properties,
-            FloatConverter: self.float_properties,
-            ColorConverter: self.color_properties,
-            ObjectPointerConverter: self.object_pointer_properties,
-            FeedbackSequenceConverter: self.feedback_sequence_properties,
-        }
-        
-        properties = properties_by_converter[converter]
-        if tag == "FileName":
-            properties = self.filename_properties
-        if replace:
-            for item in properties:
-                if item.tag == tag:
-                    item.value = value
-                    return
-        properties.add()
-        properties[-1].tag = tag
-        properties[-1].value = value
-        
-    def from_node(self, node):
-        self.tag = node.tag
-        for child_node in list(node):
-            if len(list(child_node)) == 0:
-                value = child_node.text
-                if value is None:
-                    value = ""
-                self.set(child_node.tag, value)
-            else:
-                self.dynamic_properties.add()
-                self.dynamic_properties[-1].from_node(child_node)
-        return self
-
-    def to_node(self, target_node):
-        target_node.tag = self.tag
-        if self.config_type:
-            find_or_create(target_node, "ConfigType").text = self.config_type
-        for property_group, converter in [
-                            (self.feedback_sequence_properties, FeedbackSequenceConverter),
-                            (self.string_properties, StringConverter),
-                            (self.int_properties, IntConverter),
-                            (self.filename_properties, StringConverter),
-                            (self.float_properties, FloatConverter),
-                            (self.object_pointer_properties, ObjectPointerConverter),
-                            (self.boolean_properties, BoolConverter),
-                        ]:
-            for prop in property_group:
-                value_string = converter.to_string(prop.value)
-                #It is better to always create a new subelement - otherwise there can only be one of each tag.
-                #Or does this create any problems?
-                #find_or_create(target_node, prop.tag).text = value_string
-                ET.SubElement(target_node, prop.tag).text = value_string
-        for dyn_prop in self.dynamic_properties:
-            if dyn_prop.deleted:
-                continue
-            subnode = ET.SubElement(target_node, dyn_prop.tag)
-            dyn_prop.to_node(subnode)
-        return target_node
-    
-    def draw(self, layout, split_ratio = 0.3, first_level_property = True):
-        col = layout.column()
-        header = col.row()
-        split = header.split(factor=0.6)
-        split.label(text = f"{self.tag}: {self.config_type}")
-        split.prop(self, "hidden", icon = "HIDE_OFF")
-        if not first_level_property:
-            split.prop(self, "deleted", icon = "PANEL_CLOSE")
-        if self.hidden:
-            return
-        col.separator(factor = 1.0)
-        for kw_properties in [self.feedback_sequence_properties, self.filename_properties,self.boolean_properties,
-                              self.int_properties, self.float_properties, self.string_properties, self.color_properties, self.object_pointer_properties]:
-            for item in kw_properties:
-                row = col.row()
-                split = row.split(factor=split_ratio)
-                split.alignment = "RIGHT"
-                split.label(text = item.tag)
-                split.prop(item, "value")
-        
-        for item in self.dynamic_properties:
-            if item.deleted:
-                continue
-            box = col.box()
-            item.draw(box, split_ratio, False)
-    
-
-class PT_AnnoScenePropertyPanel(Panel):
-    bl_label = "Anno Scene"
-    bl_idname = "VIEW_3D_PT_AnnoScene"
-    bl_space_type = 'VIEW_3D'
-    bl_region_type = 'UI'
-    bl_category = 'Anno Object' 
-    #bl_context = "object"
-    
-    @classmethod
-    def poll(cls, context):
-        return True
-            
-    def draw(self, context):
-        layout = self.layout
-        col = layout.column()
-        
-        col.prop(context.scene, "anno_mod_folder")
-
-class ConvertCf7DummyToDummy(Operator):
-    bl_idname = "object.convertcf7dummy"
-    bl_label = "Convert to SAFE Dummy"
-
-    def execute(self, context):
-        obj = context.active_object
-        obj.anno_object_class_str = obj.anno_object_class_str.replace("Cf7", "")
-        obj.name = obj.name.replace("Cf7", "")
-        return {'FINISHED'}
-
-class DuplicateDummy(Operator):
-    """Duplicates the dummy and gives the duplicate a higher id. Assumes a name like dummy_42."""
-    bl_idname = "object.duplicatedummy"
-    bl_label = "Duplicate Dummy"
-
-    def execute(self, context):
-        obj = context.active_object
-        duplicate = obj.copy()
-        bpy.context.scene.collection.objects.link(duplicate)
-        name = duplicate.dynamic_properties.get_string("Name")
-        head = name.rstrip('0123456789')
-        tail = name[len(head):]
-        tail = str(int(tail)+1)
-        name = head + tail
-        duplicate.dynamic_properties.set("Name", name, replace = True)
-        duplicate.name = "Dummy_" + name
-        return {'FINISHED'}
-
-
-class LoadAnimations(Operator):
-    """Loads all animations specified in the animations section of this model."""
-    bl_idname = "object.load_animations"
-    bl_label = "Load Animations"
-
-    def execute(self, context):
-        obj = context.active_object
-        node = obj.dynamic_properties.to_node(ET.Element("Config"))
-        animations_node = node.find("Animations")
-        if animations_node is not None:
-            
-            animations_container = AnimationsNode.xml_to_blender(ET.Element("Animations"), obj)
-            animations_container.name = "ANIMATIONS_"+ obj.name.replace("MODEL_", "")
-            
-            for i, anim_node in enumerate(list(animations_node)):
-                ET.SubElement(anim_node, "ModelFileName").text = get_text(node, "FileName")
-                ET.SubElement(anim_node, "AnimationIndex").text = str(i)
-                Animation.xml_to_blender(anim_node, animations_container)
-            obj.dynamic_properties.remove("Animations")
-            
-            for anim_obj in animations_container.children:
-                for armature in anim_obj.children:
-                    for anim_mesh in armature.children:
-                        for m_idx, material in enumerate(obj.data.materials):
-                            anim_mesh.data.materials[m_idx] = material
-        return {'FINISHED'}
-    
-class ShowSequence(Operator):
-    """Makes all animations belonging to this sequence visible and all others invisible. Also changes the display mode of the main models into Wireframe."""
-    bl_idname = "object.show_sequence"
-    bl_label = "Show Sequence"
-
-    
-    def set_hide_viewport_recursive(self, obj, hide):
-        obj.hide_viewport = hide
-        for o in obj.children:
-            self.set_hide_viewport_recursive(o, hide)
-
-    def show_animation(self, model_obj, animation_id):
-        for animations_obj in model_obj.children:
-            for i, anim_obj in enumerate(animations_obj.children):
-                anim_node = anim_obj.dynamic_properties.to_node(ET.Element("Config"))
-                anim_idx = get_text(anim_node, "AnimationIndex")
-                if animation_id == anim_idx:
-                    self.set_hide_viewport_recursive(anim_obj, False)
-                else:
-                    self.set_hide_viewport_recursive(anim_obj, True)
-    def show_sequence(self, seq_obj):
-        for track_obj in seq_obj.children:
-            for track_element_obj in track_obj.children:
-                track_element_node = track_element_obj.dynamic_properties.to_node(ET.Element("Track"))
-                model_name = get_text(track_element_node, "BlenderModelID", "")
-                animation_id = get_text(track_element_node, "AnimationID", "")
-                if animation_id != "" and model_name != "" and model_name in bpy.data.objects:
-                    model_obj = bpy.data.objects[model_name]
-                    self.show_animation(model_obj, animation_id)
-                    model_obj.display_type = "WIRE"
-    def get_main_file_obj(self, obj):
-        main_file_obj = obj
-        while get_anno_object_class(main_file_obj) != MainFile:
-            if main_file_obj.parent is not None:
-                main_file_obj = main_file_obj.parent
-        if get_anno_object_class(main_file_obj) == MainFile:
-            return main_file_obj
-        return None
-    
-    def show_sequences_in_subfiles(self, main_file_obj, selected_sequence_id):
-        if main_file_obj is not None:
-            for anim_sequences in main_file_obj.children:
-                if not get_anno_object_class(anim_sequences) == AnimationSequences:
-                    continue
-                for subfile_seq in anim_sequences.children:
-                    if not get_anno_object_class(subfile_seq) == AnimationSequence:
-                        continue
-                    seq_node = subfile_seq.dynamic_properties.to_node(ET.Element("Config"))
-                    sequence_id = get_text(seq_node, "SequenceID")
-                    if selected_sequence_id == sequence_id:
-                        self.show_sequence(subfile_seq)
-            for file_obj in main_file_obj.children:
-                if not get_anno_object_class(file_obj) == SubFile:
-                    continue
-                for subfile_main_file_obj in file_obj.children:
-                    if not get_anno_object_class(subfile_main_file_obj) == MainFile:
-                        continue
-                    self.show_sequences_in_subfiles(subfile_main_file_obj, selected_sequence_id)
-                    
-    def execute(self, context):
-        seq_obj = context.active_object
-        seq_node = seq_obj.dynamic_properties.to_node(ET.Element("Config"))
-        selected_sequence_id = get_text(seq_node, "SequenceID")
-        
-        #self.show_sequence(seq_obj)
-        main_file_obj = self.get_main_file_obj(seq_obj)
-        self.show_sequences_in_subfiles(main_file_obj, selected_sequence_id)
-        return {'FINISHED'}
-    
-class ShowModel(Operator):
-    """Makes all animations belonging to this sequence invisible and instead displays the normal mesh only."""
-    bl_idname = "object.show_model"
-    bl_label = "Show Model"
-
-    
-    def set_hide_viewport_recursive(self, obj, hide):
-        obj.hide_viewport = hide
-        for o in obj.children:
-            self.set_hide_viewport_recursive(o, hide)
-
-    def hide_animation(self, model_obj):
-        for animations_obj in model_obj.children:
-            for i, anim_obj in enumerate(animations_obj.children):
-                self.set_hide_viewport_recursive(anim_obj, True)
-    def hide_sequence(self, seq_obj):
-        for track_obj in seq_obj.children:
-            for track_element_obj in track_obj.children:
-                track_element_node = track_element_obj.dynamic_properties.to_node(ET.Element("Track"))
-                model_name = get_text(track_element_node, "BlenderModelID", "")
-                animation_id = get_text(track_element_node, "AnimationID", "")
-                if animation_id != "" and model_name != "" and model_name in bpy.data.objects:
-                    model_obj = bpy.data.objects[model_name]
-                    self.hide_animation(model_obj)
-                    model_obj.display_type = "TEXTURED"
-                    
-    def get_main_file_obj(self, obj):
-        main_file_obj = obj
-        while get_anno_object_class(main_file_obj) != MainFile:
-            if main_file_obj.parent is not None:
-                main_file_obj = main_file_obj.parent
-        if get_anno_object_class(main_file_obj) == MainFile:
-            return main_file_obj
-        return None
-    
-    def hide_sequences_in_subfiles(self, main_file_obj, selected_sequence_id):
-        if main_file_obj is not None:
-            for anim_sequences in main_file_obj.children:
-                if not get_anno_object_class(anim_sequences) == AnimationSequences:
-                    continue
-                for subfile_seq in anim_sequences.children:
-                    if not get_anno_object_class(subfile_seq) == AnimationSequence:
-                        continue
-                    seq_node = subfile_seq.dynamic_properties.to_node(ET.Element("Config"))
-                    sequence_id = get_text(seq_node, "SequenceID")
-                    if selected_sequence_id == sequence_id:
-                        self.hide_sequence(subfile_seq)
-            for file_obj in main_file_obj.children:
-                if not get_anno_object_class(file_obj) == SubFile:
-                    continue
-                for subfile_main_file_obj in file_obj.children:
-                    if not get_anno_object_class(subfile_main_file_obj) == MainFile:
-                        continue
-                    self.hide_sequences_in_subfiles(subfile_main_file_obj, selected_sequence_id)
-    def execute(self, context):
-        seq_obj = context.active_object
-        seq_node = seq_obj.dynamic_properties.to_node(ET.Element("Config"))
-        selected_sequence_id = get_text(seq_node, "SequenceID")
-        
-        main_file_obj = self.get_main_file_obj(seq_obj)
-        self.hide_sequences_in_subfiles(main_file_obj, selected_sequence_id)
-        return {'FINISHED'}
-
-class PT_AnnoObjectPropertyPanel(Panel):
-    bl_label = "Anno Object"
-    bl_idname = "VIEW_3D_PT_AnnoObject"
-    bl_space_type = 'VIEW_3D'
-    bl_region_type = 'UI'
-    bl_category = 'Anno Object' 
-    #bl_context = "object"
-    
-    @classmethod
-    def poll(cls, context):
-        return True
-            
-    def draw(self, context):
-        layout = self.layout
-        obj = context.active_object
-        
-        if not obj:
-            return
-        col = layout.column()
-        row = col.row()
-        
-        row.prop(obj, "anno_object_class_str")
-        row.enabled = False
-        if "Cf7" in obj.anno_object_class_str:
-            col.operator(ConvertCf7DummyToDummy.bl_idname, text = "Convert to SimpleAnnoFeedback")
-        if "Model" in obj.anno_object_class_str:
-            col.operator(LoadAnimations.bl_idname, text = "Load Animations")
-        if "AnimationSequence" == obj.anno_object_class_str:
-            col.operator(ShowSequence.bl_idname, text = "Show Sequence")    
-            col.operator(ShowModel.bl_idname, text = "Show Model")    
-            
-        if "Dummy" == obj.anno_object_class_str:
-            col.operator(DuplicateDummy.bl_idname, text = "Duplicate Dummy (ID Increment)")
-        col.prop(obj, "parent")
-        dyn = obj.dynamic_properties
-        dyn.draw(col)
-
-class PT_AnnoMaterialObjectPropertyPanel(Panel):
-    bl_label = "Anno Material"
-    bl_idname = "WINDOW_PT_AnnoMaterialObject"
-    bl_space_type = 'PROPERTIES'
-    bl_region_type = 'WINDOW'
-    bl_category = 'Anno Material' 
-    bl_context = "material"
-        
-    def draw(self, context):
-            layout = self.layout
-            obj = context.active_object.active_material
- 
-            if not obj:
-                return
-            col = layout.column()
-            dyn = obj.dynamic_properties
-            dyn.draw(col, 0.5)
-class AnnoImageTextureProperties(PropertyGroup):
-    enabled : BoolProperty( #type: ignore
-            name='Enabled',
-            description='',
-            default=True)
-    original_file_extension: EnumProperty( #type: ignore
-            name='Extension',
-            description='Some textures are stored as .png (for example default masks). Use .psd for your own textures (saved as .dds).',
-            items = [
-                (".psd", ".psd", ".psd or .dds"),
-                (".png", ".png", ".psd or .dds"),
-            ],
-            default='.psd'
-            )        
-
-class PT_AnnoImageTexture(Panel):
-    bl_label = "Anno Texture"
-    bl_idname = "SCENE_PT_AnnoTexture"
-    bl_space_type = 'NODE_EDITOR'
-    bl_region_type = 'UI'
-    bl_category = 'Anno Texture' 
-    #bl_context = "object"
-    
-    @classmethod
-    def poll(cls, context):
-        return (context.space_data.type == 'NODE_EDITOR' and
-                context.space_data.tree_type == 'ShaderNodeTree' and type(context.active_node) == bpy.types.ShaderNodeTexImage)
-
-    def draw(self, context):
-        layout = self.layout
-        node = context.active_node.anno_properties
-        col = layout.column()
-        col.prop(node, "enabled")
-        col.prop(node, "original_file_extension")
-
-class Material:
-    """
-    Can be created from an xml material node. Stores with diffuse, normal and metal texture paths and can create a corresponding blender material from them.
-    Uses a cache to avoid creating the exact same blender material multiple times when loading just one .cfg 
-    """
-    
-    texture_definitions = {
-        "cModelDiffTex":"DIFFUSE_ENABLED",
-        "cModelNormalTex":"NORMAL_ENABLED",
-        "cModelMetallicTex":"METALLIC_TEX_ENABLED",
-        "cSeparateAOTex":"SEPARATE_AO_TEXTURE",
-        "cHeightMap":"HEIGHT_MAP_ENABLED",
-        "cNightGlowMap":"NIGHT_GLOW_ENABLED", 
-        "cDyeMask":"DYE_MASK_ENABLED"
-    }
-    texture_names = {
-        "diffuse":"cModelDiffTex",
-        "normal":"cModelNormalTex",
-        "metallic":"cModelMetallicTex",
-        "ambient":"cSeparateAOTex",
-        "height":"cHeightMap",
-        "night_glow":"cNightGlowMap",
-        "dye":"cDyeMask",
-    }
-    # color_definitions = {
-    #     "cDiffuseColor":("cDiffuseColor.r", "cDiffuseColor.g", "cDiffuseColor.b"),
-    #     "cEmissiveColor":("cEmissiveColor.r", "cEmissiveColor.g", "cEmissiveColor.b"),
-    # }         "":"",
-    
-
-    color_definitions = ["cDiffuseColor", "cEmissiveColor"]
-    custom_property_default_value = {
-        "ShaderID":"", "VertexFormat":"", "NumBonesPerVertex":"", "cUseTerrainTinting":"", "Common":"", \
-        "cTexScrollSpeed":"", "cParallaxScale":"", "PARALLAX_MAPPING_ENABLED":"", \
-        "SELF_SHADOWING_ENABLED":"", "WATER_CUTOUT_ENABLED":"", "TerrainAdaption":"", "ADJUST_TO_TERRAIN_HEIGHT":"", "VERTEX_COLORED_TERRAIN_ADAPTION":"", \
-        "ABSOLUTE_TERRAIN_ADAPTION":"", "Environment":"", "cUseLocalEnvironmentBox":"", "cEnvironmentBoundingBox.x":"", "cEnvironmentBoundingBox.y":"", "cEnvironmentBoundingBox.z":"", \
-        "cEnvironmentBoundingBox.w":"", "Glow":"", "GLOW_ENABLED":"", \
-        "WindRipples":"", "WIND_RIPPLES_ENABLED":"", "cWindRippleTex":"", "cWindRippleTiling":"", "cWindRippleSpeed":"", "cWindRippleNormalIntensity":"", \
-        "cWindRippleMeshIntensity":"", "DisableReviveDistance":"", "cGlossinessFactor":"", "cOpacity":"",
-    }
-    materialCache: Dict[Tuple[Any,...], Material] = {}
-
-    def __init__(self):
-        self.textures: Dict[str, str] = {}
-        self.texture_enabled: Dict[str, bool] = {}
-        self.colors: Dict[str, List[float]] = {}
-        self.custom_properties: Dict[str, Any] = {}
-        self.name: str = "Unnamed Material"
-        self.node = None
-    @classmethod
-    def from_material_node(cls, material_node: ET.Element) -> Material:
-        instance = cls()
-        instance.name = get_text_and_delete(material_node, "Name", "Unnamed Material")
-        for texture_name, texture_enabled_flag in cls.texture_definitions.items():
-            texture_path = get_text_and_delete(material_node, texture_name)
-            instance.textures[texture_name] = texture_path
-            instance.texture_enabled[texture_name] = bool(int(get_text(material_node, texture_enabled_flag, "0")))
-        for color_name in cls.color_definitions:
-            color = [1.0, 1.0, 1.0]
-            color[0] = float(get_text_and_delete(material_node, color_name + ".r", 1.0))
-            color[1] = float(get_text_and_delete(material_node, color_name + ".g", 1.0))
-            color[2] = float(get_text_and_delete(material_node, color_name + ".b", 1.0))
-            instance.colors[color_name] = color
-        #for prop, default_value in cls.custom_property_default_value.items():
-            #value = string_to_fitting_type(get_text(material_node, prop, default_value))
-            #if value is not None:
-            #    instance.custom_properties[prop] = value
-        instance.node = material_node
-        return instance
-    
-    @classmethod
-    def from_filepaths(cls, name: str, diff_path: str, norm_path: str, metal_path: str) -> Material:
-        element = ET.fromstring(f"""
-            <Config>
-                <Name>{name}</Name>
-                <cModelDiffTex>{diff_path}</cModelDiffTex>
-                <cModelNormalTex>{metal_path}</cModelNormalTex>
-                <cModelMetallicTex>{metal_path}</cModelMetallicTex>
-            </Config>                        
-        """)
-        return cls.from_material_node(element)
-    
-    
-    @classmethod
-    def from_default(cls) -> Material:
-        element = ET.fromstring(f"""
-            <Config>
-                <Name>NEW_MATERIAL<Name>
-            </Config>                        
-        """)
-        return cls.from_material_node(element)
-         
-    @classmethod
-    def from_blender_material(cls, blender_material) -> Material:
-        instance = cls()
-        instance.node = blender_material.dynamic_properties.to_node(ET.Element("Material"))
-        instance.name = blender_material.name
-        for texture_name in cls.texture_definitions.keys():
-            shader_node = blender_material.node_tree.nodes[texture_name] #Assumes that the nodes collection allows this lookup
-            if not shader_node.image:
-                instance.textures[texture_name] = ""
-                instance.texture_enabled[texture_name] = shader_node.anno_properties.enabled
-                continue
-            filepath_full = os.path.realpath(bpy.path.abspath(shader_node.image.filepath, library=shader_node.image.library))
-            texture_path = to_data_path(filepath_full)
-            #Rename "data/.../some_diff_0.png" to "data/.../some_diff.psd"
-            extension = shader_node.anno_properties.original_file_extension
-            texture_path = Path(texture_path.as_posix().replace(instance.texture_quality_suffix()+".", ".")).with_suffix(extension)
-            instance.textures[texture_name] = texture_path.as_posix()
-            instance.texture_enabled[texture_name] = shader_node.anno_properties.enabled
-        for color_name in cls.color_definitions:
-            color = [1.0, 1.0, 1.0]
-            shader_node = blender_material.node_tree.nodes.get(color_name, None)
-            if shader_node:
-                inputs = shader_node.inputs
-                color = [inputs[0].default_value, inputs[1].default_value, inputs[2].default_value]
-            instance.colors[color_name] = color
-        for prop, default_value in cls.custom_property_default_value.items():
-            if prop not in blender_material:
-                if default_value:
-                    instance.custom_properties[prop] = default_value
-                continue
-            instance.custom_properties[prop] = blender_material[prop]
-        return instance
-    
-    def texture_quality_suffix(self):
-        return "_"+IO_AnnocfgPreferences.get_texture_quality()
-    
-    def to_xml_node(self, parent: ET.Element) -> ET.Element:
-        node = self.node
-        if not parent is None:
-            parent.append(node)
-        # node = ET.SubElement(parent, "Config")
-        #ET.SubElement(node, "ConfigType").text = "MATERIAL"
-        ET.SubElement(node, "Name").text = self.name
-        for texture_name in self.texture_definitions.keys():
-            texture_path = self.textures[texture_name]
-            if texture_path != "":
-                ET.SubElement(node, texture_name).text = texture_path
-        for color_name in self.color_definitions:
-            ET.SubElement(node, color_name + ".r").text = format_float(self.colors[color_name][0])
-            ET.SubElement(node, color_name + ".g").text = format_float(self.colors[color_name][1])
-            ET.SubElement(node, color_name + ".b").text = format_float(self.colors[color_name][2])
-        for texture_name, texture_enabled_flag in self.texture_definitions.items():
-            used_value = self.texture_enabled[texture_name]
-            find_or_create(node, texture_enabled_flag).text = str(int(used_value))
-        for prop, value in self.custom_properties.items():
-            if value == "":
-                continue
-            if type(value) == float:
-                value = format_float(value)
-            ET.SubElement(node, prop).text = str(value)
-        return node
-    
-    def convert_to_png(self, fullpath: Path) -> bool:
-        """Converts the .dds file to .png. Returns True if successful, False otherwise.
-
-        Args:
-            fullpath (str): .dds file
-
-        Returns:
-            bool: Successful
-        """
-        if not IO_AnnocfgPreferences.get_path_to_texconv().exists():
-            return False
-        if not fullpath.exists():
-            return False
-        try:
-            subprocess.call(f"\"{IO_AnnocfgPreferences.get_path_to_texconv()}\" -ft PNG -sepalpha -y -o \"{fullpath.parent}\" \"{fullpath}\"")
-        except:
-            return False
-        return fullpath.with_suffix(".png").exists()
-    
-    def get_texture(self, texture_path: Path):
-        """Tries to find the texture texture_path with ending "_0.png" (quality setting can be changed) in the list of loaded textures.
-        Otherwise loads it. If it is not existing but the corresponding .dds exists, converts it first.
-
-        Args:
-            texture_path (str): f.e. "data/.../texture_diffuse.psd"
-
-        Returns:
-            [type]: The texture or None.
-        """
-        if texture_path == Path(""):
-            return None
-        texture_path = Path(texture_path)
-        texture_path = Path(texture_path.parent, texture_path.stem + self.texture_quality_suffix()+".dds")
-        png_file = texture_path.with_suffix(".png")
-        image = bpy.data.images.get(str(png_file.name), None)
-        if image is not None:
-            return image
-        fullpath = data_path_to_absolute_path(texture_path)
-        png_fullpath = data_path_to_absolute_path(png_file)
-        if not png_fullpath.exists():
-            success = self.convert_to_png(fullpath)
-            if not success:
-                print("Failed to convert texture", fullpath)
-                return None
-        image = bpy.data.images.load(str(png_fullpath))
-        return image
-
-    
-
-    def get_material_cache_key(self):
-        attribute_list = tuple([self.name] + list(self.textures.items()) + list([(a, tuple(b)) for a, b in self.colors.items()]) + list(self.custom_properties.items()))
-        return hash(attribute_list)
-    
-    def create_anno_shader(self):
-        anno_shader = bpy.data.node_groups.new('AnnoShader', 'ShaderNodeTree')
-        
-        anno_shader.inputs.new("NodeSocketColor", "cDiffuse")
-        anno_shader.inputs.new("NodeSocketColor", "cDiffuseMultiplier")
-        anno_shader.inputs.new("NodeSocketFloat", "Alpha")
-        anno_shader.inputs.new("NodeSocketColor", "cNormal")
-        anno_shader.inputs.new("NodeSocketFloat", "Glossiness")
-        anno_shader.inputs.new("NodeSocketColor", "cMetallic")
-        anno_shader.inputs.new("NodeSocketColor", "cHeight")
-        anno_shader.inputs.new("NodeSocketColor", "cNightGlow")
-        anno_shader.inputs.new("NodeSocketColor", "cEmissiveColor")
-        anno_shader.inputs.new("NodeSocketFloat", "EmissionStrength")
-        anno_shader.inputs.new("NodeSocketColor", "cDyeMask")
-        
-        
-        anno_shader.outputs.new("NodeSocketShader", "Shader")
-        
-        inputs = self.add_shader_node(anno_shader, "NodeGroupInput", 
-                                        position = (0, 0), 
-                                    ).outputs
-        mix_c_diffuse = self.add_shader_node(anno_shader, "ShaderNodeMixRGB",
-                                        position = (1, 4),
-                                        default_inputs = {
-                                            0 : 1.0,
-                                        },
-                                        inputs = {
-                                            "Color1" : inputs["cDiffuseMultiplier"],
-                                            "Color2" : inputs["cDiffuse"],
-                                        },
-                                        blend_type = "MULTIPLY",
-                                    )
-        dye_mask = self.add_shader_node(anno_shader, "ShaderNodeRGBToBW",
-                                        position = (1, 3),
-                                        inputs = {
-                                            "Color" : inputs["cDyeMask"],
-                                        },
-                                    )
-        final_diffuse = self.add_shader_node(anno_shader, "ShaderNodeMixRGB",
-                                        position = (2, 3),
-                                        default_inputs = {
-                                            "Color2" : (1.0, 0.0, 0.0, 1.0),
-                                        },
-                                        inputs = {
-                                            "Fac" : dye_mask.outputs["Val"],
-                                            "Color1" : mix_c_diffuse.outputs["Color"],
-                                        },
-                                        blend_type = "MULTIPLY",
-                                    )
-        #Normals
-        separate_normal = self.add_shader_node(anno_shader, "ShaderNodeSeparateRGB",
-                                        position = (1, 2),
-                                        inputs = {
-                                            "Image" : inputs["cNormal"],
-                                        },
-                                    )
-        #Calc normal blue
-        square_x = self.add_shader_node(anno_shader, "ShaderNodeMath",
-                                        position = (2, 1.5),
-                                        operation = "POWER",
-                                        inputs = {
-                                            0 : separate_normal.outputs["R"],
-                                        },
-                                        default_inputs = {
-                                            1 : 2.0
-                                        },
-                                    )
-        square_y = self.add_shader_node(anno_shader, "ShaderNodeMath",
-                                        position = (2, 2.5),
-                                        operation = "POWER",
-                                        inputs = {
-                                            0 : separate_normal.outputs["G"],
-                                        },
-                                        default_inputs = {
-                                            1 : 2.0
-                                        },
-                                    )
-        add_squares = self.add_shader_node(anno_shader, "ShaderNodeMath",
-                                        position = (2.5, 2),
-                                        operation = "ADD",
-                                        inputs = {
-                                            0 : square_x.outputs["Value"],
-                                            1 : square_y.outputs["Value"],
-                                        },
-                                    )
-        inverted_add_squares = self.add_shader_node(anno_shader, "ShaderNodeMath",
-                                        position = (3, 2),
-                                        operation = "SUBTRACT",
-                                        inputs = {
-                                            1 : add_squares.outputs["Value"],
-                                        },
-                                        default_inputs = {
-                                            0 : 1.0
-                                        },
-                                    )
-        normal_blue = self.add_shader_node(anno_shader, "ShaderNodeMath",
-                                        position = (3.5, 2),
-                                        operation = "SQRT",
-                                        inputs = {
-                                            0 : inverted_add_squares.outputs["Value"],
-                                        },
-                                    )
-        
-        combine_normal = self.add_shader_node(anno_shader, "ShaderNodeCombineRGB",
-                                        position = (4, 2),
-                                        inputs = {
-                                            "R" : separate_normal.outputs["R"],
-                                            "G" : separate_normal.outputs["G"],
-                                            "B" : normal_blue.outputs["Value"],
-                                        },
-                                    )
-        normal_map = self.add_shader_node(anno_shader, "ShaderNodeNormalMap",
-                                        position = (5, 2),
-                                        default_inputs = {
-                                            0 : 0.5,
-                                        },
-                                        inputs = {
-                                            "Color" : combine_normal.outputs["Image"],
-                                        },
-                                    )
-        height_bw = self.add_shader_node(anno_shader, "ShaderNodeRGBToBW",
-                                        position = (5, 3),
-                                        inputs = {
-                                            "Color" : inputs["cHeight"],
-                                        },
-                                    )
-        bump_map = self.add_shader_node(anno_shader, "ShaderNodeBump",
-                                        position = (6, 2),
-                                        default_inputs = {
-                                            0 : 0.5,
-                                        },
-                                        inputs = {
-                                            "Height" : height_bw.outputs["Val"],
-                                            "Normal" : normal_map.outputs["Normal"],
-                                        },
-                                    )
-        #Roughness
-        roughness = self.add_shader_node(anno_shader, "ShaderNodeMath",
-                                position = (3, 0),
-                                operation = "SUBTRACT",
-                                inputs = {
-                                    1 : inputs["Glossiness"],
-                                },
-                                default_inputs = {
-                                    0 : 1.0
-                                },
-                            )
-        #Metallic
-        metallic = self.add_shader_node(anno_shader, "ShaderNodeRGBToBW",
-                                        position = (1, 3),
-                                        inputs = {
-                                            "Color" : inputs["cMetallic"],
-                                        },
-                                    )
-        #Emission
-        scaled_emissive_color = self.add_shader_node(anno_shader, "ShaderNodeVectorMath",         
-                            operation = "SCALE",
-                            name = "EmissionScale",
-                            position = (1, -1),
-                            default_inputs = {
-                                "Scale": 10,
-                            },
-                            inputs = {
-                                "Vector" : inputs["cEmissiveColor"],
-                            }
-        )
-        combined_emissive_color = self.add_shader_node(anno_shader, "ShaderNodeVectorMath",         
-                            operation = "MULTIPLY",
-                            position = (2, -1),
-                            inputs = {
-                                0 : final_diffuse.outputs["Color"],
-                                1 : scaled_emissive_color.outputs["Vector"],
-                            }
-        )
-        object_info = self.add_shader_node(anno_shader, "ShaderNodeObjectInfo",         
-                            position = (1, -2),
-        )
-        random_0_1 = self.add_shader_node(anno_shader, "ShaderNodeMath",  
-                            operation = "FRACT",   
-                            position = (2, -2),
-                            inputs = {
-                                "Value" : object_info.outputs["Location"],
-                            }
-        )
-        color_ramp_node = self.add_shader_node(anno_shader, "ShaderNodeValToRGB",  
-                            position = (3, -2),
-                            inputs = {
-                                "Fac" : random_0_1.outputs["Value"],
-                            }
-        )
-
-        color_ramp = color_ramp_node.color_ramp
-        color_ramp.elements[0].color = (1.0, 0.0, 0.0,1)
-        color_ramp.elements[1].position = (2.0/3.0)
-        color_ramp.elements[1].color = (0.0, 0.0, 1.0,1)
-        
-        color_ramp.elements.new(1.0/3.0)
-        color_ramp.elements[1].color = (0.0, 1.0, 0.0,1)
-        color_ramp.interpolation = "CONSTANT"
-        
-        location_masked_emission = self.add_shader_node(anno_shader, "ShaderNodeVectorMath",         
-                            operation = "MULTIPLY",
-                            position = (4, -2),
-                            inputs = {
-                                0 : color_ramp_node.outputs["Color"],
-                                1 : inputs["cNightGlow"],
-                            }
-        )
-        
-        final_emission_color = self.add_shader_node(anno_shader, "ShaderNodeMixRGB",         
-                            blend_type = "MIX",
-                            position = (5, -1),
-                            default_inputs = {
-                                "Color1" : (0.0, 0.0 ,0.0, 1.0)
-                            },
-                            inputs = {
-                                "Fac" : location_masked_emission.outputs["Vector"],
-                                "Color2" : combined_emissive_color.outputs["Vector"],
-                            }
-        )
-        
-        bsdf = self.add_shader_node(anno_shader, "ShaderNodeBsdfPrincipled", 
-                                        position = (4, 0), 
-                                        inputs = {
-                                            "Alpha" : inputs["Alpha"],
-                                            "Roughness" : roughness.outputs["Value"],
-                                            "Normal" : bump_map.outputs["Normal"],
-                                            "Base Color" : final_diffuse.outputs["Color"],
-                                            "Metallic" : metallic.outputs["Val"],
-                                            "Emission Strength" : inputs["EmissionStrength"],
-                                            "Emission" : final_emission_color.outputs["Color"],
-                                            
-                                            
-                                        },
-                                    )
-        outputs = self.add_shader_node(anno_shader, "NodeGroupOutput", 
-                                        position = (5, 0), 
-                                        inputs = {
-                                            "Shader" : bsdf.outputs["BSDF"]
-                                        },
-                                    )
-
-    
-    def add_anno_shader(self, nodes):
-        group = nodes.new(type='ShaderNodeGroup')
-        if not "AnnoShader" in bpy.data.node_groups:
-            self.create_anno_shader()            
-        group.node_tree = bpy.data.node_groups["AnnoShader"]
-        return group
-        
-    def as_blender_material(self):
-
-        if self.get_material_cache_key() in Material.materialCache:
-            return Material.materialCache[self.get_material_cache_key()]
-        
-        material = bpy.data.materials.new(name=self.name)
-        
-        material.dynamic_properties.from_node(self.node)
-        material.use_nodes = True
-        
-        positioning_unit = (300, 300)
-        positioning_offset = (0, 3 * positioning_unit[1])
-        
-        
-        for i, texture_name in enumerate(self.texture_definitions.keys()):
-            texture_node = material.node_tree.nodes.new('ShaderNodeTexImage')
-            texture_path = Path(self.textures[texture_name])
-            texture = self.get_texture(texture_path)
-            if texture is not None:
-                texture_node.image = texture
-                if "Norm" in texture_name or "Metal" in texture_name or "Height" in texture_name:
-                    texture_node.image.colorspace_settings.name = 'Non-Color'
-            texture_node.name = texture_name
-            texture_node.label = texture_name
-            texture_node.location.x -= 4 * positioning_unit[0] - positioning_offset[0]
-            texture_node.location.y -= i * positioning_unit[1] - positioning_offset[1]
-
-            texture_node.anno_properties.enabled = self.texture_enabled[texture_name]
-            extension = texture_path.suffix
-            if extension not in [".png", ".psd"]:
-                if texture_path != Path(""):
-                    print("Warning: Unsupported texture file extension", extension, texture_path)
-                extension = ".psd"
-            texture_node.anno_properties.original_file_extension = extension
-        
-        node_tree = material.node_tree
-        links = node_tree.links
-        nodes = node_tree.nodes
-        
-        anno_shader = self.add_anno_shader(nodes)
-        material.node_tree.nodes.remove(nodes["Principled BSDF"])
-        
-        emissive_color = self.add_shader_node(node_tree, "ShaderNodeCombineRGB",
-                            name = "cEmissiveColor",
-                            position = (3, 6.5),
-                            default_inputs = {
-                                "R": self.colors["cEmissiveColor"][0],
-                                "G": self.colors["cEmissiveColor"][1],
-                                "B": self.colors["cEmissiveColor"][2],
-                            },
-                            inputs = {}
-        )
-        c_diffuse_mult = self.add_shader_node(node_tree, "ShaderNodeCombineRGB",
-                            name = "cDiffuseColor",
-                            position = (2, 6.5),
-                            default_inputs = {
-                                "R": self.colors["cDiffuseColor"][0],
-                                "G": self.colors["cDiffuseColor"][1],
-                                "B": self.colors["cDiffuseColor"][2],
-                            },
-                            inputs = {}
-        )
-        
-        links.new(anno_shader.inputs["cDiffuse"], nodes[self.texture_names["diffuse"]].outputs[0])
-        links.new(anno_shader.inputs["cNormal"], nodes[self.texture_names["normal"]].outputs[0])
-        links.new(anno_shader.inputs["cMetallic"], nodes[self.texture_names["metallic"]].outputs[0])
-        links.new(anno_shader.inputs["cHeight"], nodes[self.texture_names["height"]].outputs[0])
-        links.new(anno_shader.inputs["cNightGlow"], nodes[self.texture_names["night_glow"]].outputs[0])
-        links.new(anno_shader.inputs["cDyeMask"], nodes[self.texture_names["dye"]].outputs[0])
-        
-        links.new(anno_shader.inputs["cDiffuseMultiplier"], c_diffuse_mult.outputs[0])
-        links.new(anno_shader.inputs["cEmissiveColor"], emissive_color.outputs[0])
-        
-        links.new(anno_shader.inputs["Alpha"], nodes[self.texture_names["diffuse"]].outputs["Alpha"])
-        links.new(anno_shader.inputs["Glossiness"], nodes[self.texture_names["normal"]].outputs["Alpha"])
-        
-        
-        links.new(nodes["Material Output"].inputs["Surface"], anno_shader.outputs["Shader"])
-        
-        
-        
-        material.blend_method = "CLIP"
-
-        
-        #Store all kinds of properties for export
-        for prop, value in self.custom_properties.items():
-            material[prop] = value
-
-
-        Material.materialCache[self.get_material_cache_key()] = material
-        return material
-    
-    def add_shader_node(self, node_tree, node_type, **kwargs):
-        node = node_tree.nodes.new(node_type)
-        positioning_unit = (300, 300)
-        positioning_offset = (0, 3 * positioning_unit[1])
-        x,y = kwargs.pop("position", (0,0))
-        node.location.x = x* positioning_unit[0] - positioning_offset[0]
-        node.location.y = y* positioning_unit[1] - positioning_offset[1]
-        if "name" in kwargs and not "label" in kwargs:
-            kwargs["label"] = kwargs["name"]
-        for input_key, default_value in kwargs.pop("default_inputs", {}).items():
-            node.inputs[input_key].default_value = default_value
-        for input_key, input_connector in kwargs.pop("inputs", {}).items():
-             node_tree.links.new(node.inputs[input_key], input_connector)
-        for attr, value in kwargs.items():
-            setattr(node, attr, value)
-        return node
-    
-    def add_shader_node_to_material(self, material, node_type, **kwargs):
-        nodes = material.node_tree
-        return self.add_shader_node(nodes, node_type, **kwargs)
-###################################################################################################################
-
-class ClothMaterial(Material):
-    texture_definitions = {
-        "cClothDiffuseTex":"DIFFUSE_ENABLED",
-        "cClothNormalTex":"NORMAL_ENABLED",
-        "cClothMetallicTex":"METALLIC_TEX_ENABLED",
-        "cSeparateAOTex":"SEPARATE_AO_TEXTURE",
-        "cHeightMap":"HEIGHT_MAP_ENABLED",
-        "cNightGlowMap":"NIGHT_GLOW_ENABLED", 
-        "cClothDyeMask":"DYE_MASK_ENABLED"
-    }
-    texture_names = {
-        "diffuse":"cClothDiffuseTex",
-        "normal":"cClothNormalTex",
-        "metallic":"cClothMetallicTex",
-        "ambient":"cSeparateAOTex",
-        "height":"cHeightMap",
-        "night_glow":"cNightGlowMap",
-        "dye":"cClothDyeMask",
-    }
-
-def is_type(T: type, s: str) -> bool:
-    try:
-        T(s)
-        return True
-    except:
-        return False
-
-def string_to_fitting_type(s: str):
-    if is_type(int, s) and s.isnumeric():
-        return int(s)
-    if is_type(float, s):
-        return float(s)
-    return s
-    
-def get_first_or_none(list):
-    if list:
-        return list[0]
-    return None
-
-
-def get_text(node: ET.Element, query: str, default_value = "") -> str:
-    if node.find(query) is None:
-        return str(default_value)
-    if node.find(query).text is None:
-        return str(default_value)
-    return node.find(query).text
-
-def get_text_and_delete(node: ET.Element, query: str, default_value = "") -> str:
-    if node.find(query) is None:
-        return str(default_value)
-    subnode = node.find(query)
-    parent = node
-    if "/" in query:
-        query = query.rsplit("/", maxsplit=1)[0]
-        parent = node.find(query)
-    parent.remove(subnode)
-    while len(list(parent)) == 0 and parent != node:
-        elements = query.rsplit("/", maxsplit=1)
-        query = elements[0]
-        parents_parent = node
-        if len(elements) > 1:
-            parents_parent = node.find(query)
-        parents_parent.remove(parent)
-        parent = parents_parent
-    if subnode.text is None:
-        return default_value
-    return subnode.text
-
-def format_float(value: Union[float, int]):
-    return "{:.6f}".format(value)
-
-def find_or_create(parent: ET.Element, simple_query: str) -> ET.Element:
-    """Finds or creates the subnode corresponding to the simple query.
-    
-
-    Args:
-        parent (ET.Element): root node
-        simple_query (str): Only supports queries like "Config[ConfigType="ORIENTATION_TRANSFORM"]/Position/x", no advanced xpath.
-
-    Returns:
-        ET.Element: The node.
-    """
-    parts = simple_query.split("/", maxsplit = 1)
-    query = parts[0]
-    queried_node = parent.find(query)
-    if queried_node is None:
-        tag = query.split("[")[0]
-        queried_node = ET.SubElement(parent, tag)
-        if "[" in query:
-            condition = query.split("[")[1].replace("]", "")
-            subnode_tag = condition.split("=")[0].strip()
-            subnode_value = condition.split("=")[1].strip().replace('"', '').replace("'", "")
-            ET.SubElement(queried_node, subnode_tag).text = subnode_value
-    if len(parts) > 1:
-        return find_or_create(queried_node, parts[1])
-    return queried_node
 
 def convert_to_glb(fullpath: Path):
     rdm4_path = IO_AnnocfgPreferences.get_path_to_rdm4()
@@ -2067,7 +624,7 @@ class Prop(AnnoObject):
     
     prop_data_by_filename: Dict[str, Tuple[Optional[str], Optional[Material]]] = {} #avoids opening the same .prp file multiple times
     
-    prop_obj_blueprints = {} #used to copy mesh prop data 
+    prop_obj_blueprints: Dict[str, BlenderObject] = {} #used to copy mesh prop data 
     
     @classmethod
     def get_prop_data(cls, prop_filename: str) -> Tuple[Optional[str], Optional[Material]]:
@@ -2352,10 +909,8 @@ class IfoMeshHeightmap(AnnoObject):
             for b in range(width):
                 verts.append((startx + b * stepx, starty + a * stepy, heightdata[i]))
                 i += 1
-        edges = []
-        faces = []
 
-        mesh.from_pydata(verts, edges, faces)
+        mesh.from_pydata(verts, [], [])
         for i, vert in enumerate(obj.data.vertices):
             vert.co.y *= -1
             
@@ -2410,7 +965,7 @@ class Dummy(AnnoObject):
         return file_obj
     @classmethod
     def default_node(cls: Type[T]):
-        node = super().default_node()
+        node = super().default_node() # type: ignore
         node.tag = "Dummy"
         ET.SubElement(node, "Name")
         ET.SubElement(node, "HeightAdaptationMode").text = "1"
@@ -2431,7 +986,7 @@ class DummyGroup(AnnoObject):
     has_name = False
     @classmethod
     def default_node(cls: Type[T]):
-        node = super().default_node()
+        node = super().default_node() # type: ignore
         node.tag = "DummyGroup"
         ET.SubElement(node, "Name")
         return node
@@ -2472,7 +1027,7 @@ class FeedbackConfig(AnnoObject):
 
     @classmethod
     def default_node(cls: Type[T]):
-        node = super().default_node()
+        node = super().default_node() # type: ignore
         node.tag = "FeedbackConfig"
         ET.SubElement(node, "GUIDVariationList")
         ET.SubElement(node, "SequenceElements")
@@ -3088,8 +1643,8 @@ class GameObject:
         mesh_node = node.find("Mesh")
     
         location = [float(s) for s in get_text_and_delete(node, "Position", "0,0 0,0 0,0").replace(",", ".").split(" ")]
-        rotation = [1,0,0,0] 
-        scale    = [1,1,1]
+        rotation = [1.0, 0.0, 0.0, 0.0] 
+        scale    = [1.0, 1.0, 1.0]
         if mesh_node is not None:
             rotation = [float(s) for s in get_text_and_delete(mesh_node, "Orientation", "1,111 0,0 0,0 0,0").replace(",", ".").split(" ")]
             rotation = [rotation[3], rotation[0], rotation[1], rotation[2]] #xzyw -> wxzy
@@ -3203,6 +1758,8 @@ anno_object_classes = [
     IslandGamedataFile, GameObject, AnimationsNode, Animation, AnimationSequences, AnimationSequence, Track, TrackElement, IfoMeshHeightmap,
 ]
 
+def str_to_class(classname):
+    return getattr(sys.modules[__name__], classname)
 
 def get_anno_object_class(obj) -> type:
     return str_to_class(obj.anno_object_class_str)
@@ -3210,46 +1767,18 @@ def get_anno_object_class(obj) -> type:
 def set_anno_object_class(obj, cls: type):
      obj.anno_object_class_str = cls.__name__
     
-classes = [
-    AnnoImageTextureProperties,
-    PT_AnnoImageTexture,
     
-    BoolPropertyGroup,
-    IntPropertyGroup,
-    StringPropertyGroup,
-    FloatPropertyGroup,
-    FilenamePropertyGroup,
-    ColorPropertyGroup,
-    FeedbackSequencePropertyGroup,
-    ObjectPointerPropertyGroup,
     
-    PT_AnnoScenePropertyPanel,
     
-    PT_AnnoMaterialObjectPropertyPanel,
-    ConvertCf7DummyToDummy,
-    LoadAnimations,
-    DuplicateDummy,
-    ShowSequence,
-    ShowModel,
     
-    XMLPropertyGroup,
-    PT_AnnoObjectPropertyPanel,
-]
+
+    
 def register():
-    for cls in classes:
-        bpy.utils.register_class(cls)
-    bpy.types.ShaderNodeTexImage.anno_properties = bpy.props.PointerProperty(type=AnnoImageTextureProperties)
     bpy.types.Object.anno_object_class_str = bpy.props.EnumProperty(name="Anno Object Class", description = "Determines the type of the object.",
                                                                 items = [(cls.__name__, cls.__name__, cls.__name__) for cls in anno_object_classes]
                                                                 , default = "NoAnnoObject")
-    bpy.types.Object.dynamic_properties = bpy.props.PointerProperty(type = XMLPropertyGroup)
-    bpy.types.Material.dynamic_properties = bpy.props.PointerProperty(type = XMLPropertyGroup)
     #CollectionProperty(type = AnnoImageTextureProperties)
 
 def unregister():
-    del bpy.types.ShaderNodeTexImage.anno_properties
-    del bpy.types.Object.dynamic_properties
     del bpy.types.Object.anno_object_class_str
-    for cls in classes:
-        bpy.utils.unregister_class(cls)
